Check permissions on notification subscription
This commit is janky. That metadeploy.consumers should need to know
about and pull data from metadeploy.api.models does not please me. This
also can be cleaned up when some more things in the start-install branch
get merged into here. The only model that currently has any sort of
access control is Job, but that access control is described on the
viewset, the Job model instance, and possibly elsewhere. This calls for
cleanup.