Revert cache, replace deprecated command
The cache-to input doesn't work how I expect, so I need to read the docs
to see whether it's even possible to get the behavior I want. Alas.

However, scouting rule says that, since `set-output` has been deprecated
for a while, we should replace it with the environment file version.