Add omniscript-lwc-compiler as dependency (#3518)
Required to support [local compilation] of FlexCards in Vlocity CLI.
This requires the addition of a private NPM repository to allow the
installation of the compiler via:

```console
yarn add @omniscript/omniscript-lwc-compiler
```

There are alternative routes specified in the Vlocity documentation, but
users reported they were unsuccessful. The addition of a `.npmrc` file to configure the repository (see [heroku docs]) relies on an environment variable named `OMNIOUT_TOKEN` that contains the authorization token received from support.

TODO: Migrate to Yarn 2

[heroku docs]: https://devcenter.heroku.com/articles/nodejs-support#private-dependencies
[local compilation]: https://github.com/vlocityinc/vlocity_build#initial-support-for-omniscript--flexcards-local-compilation

We initially hit a build error because OMNIOUT_TOKEN  the env var isn't being
set as expected. We found that the inputs for our action must be provided in
each job, and that adding it as an environment variable is insufficent. We
updated our local build action to pass this value as an input to the build.


In addition, we replaced `set-output` with its environment file equivalent since it has been deprecated.